+++ changelog.md
@@ -1,1 +1,3 @@
 - Updated ConfigBuilder
+- Removed `show_shulker_box_items` config option
+- Added `shulker_box_item_display_type` config option